scripts/batch-ingestion/pyproject.toml
[tool.flake8]
max-line-length = 120
ignore = ["BLK999", "W503", "PT009"]

[tool.black]
line-length = 120
target-version = ["py38"]

[tool.ruff]
line-length = 120
lint.select = ["E4", "E7", "E9", "F", "B", "E501"]
lint.extend-select = ["I"]

[tool.ruff.format]
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
addopts = "-ra -v --strict-markers -m 'not manual_test'"
markers = [
    "unit_test: Tests the behavior of one function (unit) in isolation from external infrastructure.",
    "integration_test: Tests the behavior of one function that integrates with external infrastructure.",
    "end_to_end_test: Tests the behavior of a full application, end-to-end.",
    "manual_test: Marks a test as a manual test that should not be run automatically."
]
testpaths = [
    "tests"
]

[tool.coverage.run]
source = [
    "simulations_python_aggregator"
]

[tool.poetry]
name = "batch-ingestion"
version = "0.1.0"
description = "A small Python package for batch ingestion to analyze some tennis matches"
authors = ["John Larkin <john@johnjlarkin.com>"]
readme = "README.md"
packages = [{include = "data_pipeline"}]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
supabase = "^2.9.0"
sportradar-tennis-v3 = {path = "sportradar-tennis-v3"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"
mypy = "^1.11.2"
pytest = "^8.3.3"
datamodel-code-generator = "^0.26.1"
openapi-python-client = "^0.21.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


